Refactor icl loader

+++ src/utils/annotations.py
@@ -3,7 +3,7 @@ import cv2
 
 
 def draw_polygone(image, plane):
-    contours = np.array(plane.points)
+    contours = np.array(plane.points).astype(dtype=np.float)
     color_tuple = tuple([int(part) for part in plane.color])
     cv2.fillPoly(image, pts=np.int32([contours]), color=color_tuple)
     return image
